+++ lib/service/Health.dart
@@ -423,8 +423,9 @@ class Health with Service implements NotificationsListener {
       return rules?.symptoms?.groups;
     }
     else {
-      String url = "${Config().health2Url}/symptoms/symptoms.json";
-      Response response = await Network().get(url);
+      String url = "${Config().healthUrl}/covid19/symptoms";
+      String appVersion = AppVersion.majorVersion(Config().appVersion, 2);
+      Response response = await Network().get(url, auth: NetworkAuth.App, headers: { Network.RokwireVersion : appVersion });
       String responseBody = (response?.statusCode == 200) ? response.body : null;
 //TMP:String responseBody = await rootBundle.loadString('assets/sample.health.symptoms.json');
       List<dynamic> responseJson = (responseBody != null) ? AppJson.decodeList(responseBody) : null;
@@ -1328,8 +1329,9 @@ class Health with Service implements NotificationsListener {
   // Consolidated Rules
 
   Future<HealthRulesSet2> _loadRules2({String countyId}) async {
-    String url = "${Config().health2Url}/rules/county/$countyId/rules.json";
-    Response response = await Network().get(url);
+    String appVersion = AppVersion.majorVersion(Config().appVersion, 2);
+    String url = "${Config().healthUrl}/covid19/crules/county/$countyId";
+    Response response = await Network().get(url, auth: NetworkAuth.App, headers: { Network.RokwireVersion : appVersion });
     String responseBody = (response?.statusCode == 200) ? response.body : null;
 //TMP:String responseBody = await rootBundle.loadString('assets/sample.health.rules.json');
     Map<String, dynamic> responseJson = (responseBody != null) ? AppJson.decodeMap(responseBody) : null;