+++ lib/ui/health/Covid19InfoCenterPanel.dart
@@ -19,10 +19,12 @@ import 'dart:collection';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
+import 'package:flutter_html/style.dart';
 import 'package:illinois/model/Health.dart';
 import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/service/AppDateTime.dart';
 import 'package:illinois/service/Config.dart';
+import 'package:illinois/service/Connectivity.dart';
 import 'package:illinois/service/Health.dart';
 import 'package:illinois/service/Localization.dart';
 import 'package:illinois/service/NotificationService.dart';
@@ -328,7 +330,7 @@ class _Covid19InfoCenterPanelState extends State<Covid19InfoCenterPanel> impleme
     if (AppString.isStringNotEmpty(nextStepHtml)) {
       content.addAll(<Widget>[
           Container(height: 12,),
-          Html(data: nextStepHtml, onLinkTap: (url) => _onTapLink(url), defaultTextStyle: TextStyle(fontSize: 16, fontFamily: Styles().fontFamilies.regular, color: Styles().colors.textBackground),),
+          Html(data: nextStepHtml, onLinkTap: (url) => _onTapLink(url), style:{ 'body' : Style(fontFamily: Styles().fontFamilies.regular, color: Styles().colors.textBackground)},),
       ]);
     }
 
@@ -593,54 +595,90 @@ class _Covid19InfoCenterPanelState extends State<Covid19InfoCenterPanel> impleme
       ),
     );
   }
-  
+
   void _onTapCountryGuidelines() {
-    Analytics.instance.logSelect(target: "COVID-19 County Guidlines");
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => Covid19GuidelinesPanel(status: _status)));
+    if(Connectivity().isNotOffline) {
+      Analytics.instance.logSelect(target: "COVID-19 County Guidlines");
+      Navigator.push(context, CupertinoPageRoute(builder: (context) => Covid19GuidelinesPanel(status: _status)));
+    } else{
+      AppAlert.showOfflineMessage(context);
+    }
   }
 
   void _onTapCareTeam() {
-    Analytics.instance.logSelect(target: "Your Care Team");
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => Covid19CareTeamPanel(status: _status,)));
+    if(Connectivity().isNotOffline) {
+      Analytics.instance.logSelect(target: "Your Care Team");
+      Navigator.push(context, CupertinoPageRoute(builder: (context) => Covid19CareTeamPanel(status: _status,)));
+    } else{
+      AppAlert.showOfflineMessage(context);
+    }
   }
 
   void _onTapReportTest(){
-    Analytics.instance.logSelect(target: "COVID-19 Report Test");
-    Navigator.push(context, CupertinoPageRoute(builder: (context)=>Covid19AddTestResultPanel()));
+    if(Connectivity().isNotOffline) {
+      Analytics.instance.logSelect(target: "COVID-19 Report Test");
+      Navigator.push(context, CupertinoPageRoute(builder: (context) => Covid19AddTestResultPanel()));
+    } else{
+      AppAlert.showOfflineMessage(context);
+    }
   }
 
   void _onTapTestHistory(){
-    Analytics.instance.logSelect(target: "COVID-19 Test History");
-    Navigator.push(context, CupertinoPageRoute(builder: (context)=>Covid19HistoryPanel()));
+    if(Connectivity().isNotOffline) {
+      Analytics.instance.logSelect(target: "COVID-19 Test History");
+      Navigator.push(context, CupertinoPageRoute(builder: (context) => Covid19HistoryPanel()));
+    } else{
+      AppAlert.showOfflineMessage(context);
+    }
   }
 
   void _onTapFindLocations(){
-    Analytics.instance.logSelect(target: "COVID-19 Find Test Locations");
-    Navigator.push(context, CupertinoPageRoute(builder: (context)=>Covid19TestLocationsPanel()));
+    if(Connectivity().isNotOffline) {
+      Analytics.instance.logSelect(target: "COVID-19 Find Test Locations");
+      Navigator.push(context, CupertinoPageRoute(builder: (context) => Covid19TestLocationsPanel()));
+    } else{
+      AppAlert.showOfflineMessage(context);
+    }
   }
 
   void _onTapShowStatusCard(){
-    Analytics.instance.logSelect(target: "Show Status Card");
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => Covid19StatusPanel()));
+    if(Connectivity().isNotOffline) {
+      Analytics.instance.logSelect(target: "Show Status Card");
+      Navigator.push(context, CupertinoPageRoute(builder: (context) => Covid19StatusPanel()));
+    } else{
+      AppAlert.showOfflineMessage(context);
+    }
   }
 
   void _onTapSymptomCheckIn() {
-    Analytics.instance.logSelect(target: "Symptom Check-in");
-    Navigator.push(context, CupertinoPageRoute(builder: (context)=>Covid19SymptomsPanel()));
+    if(Connectivity().isNotOffline) {
+      Analytics.instance.logSelect(target: "Symptom Check-in");
+      Navigator.push(context, CupertinoPageRoute(builder: (context) => Covid19SymptomsPanel()));
+    } else{
+      AppAlert.showOfflineMessage(context);
+    }
   }
 
   void _onTapCovidWellnessCenter(){
-    Analytics.instance.logSelect(target: "Wellness Center");
-    Navigator.push(context, CupertinoPageRoute(builder: (context)=>Covid19WellnessCenter()));
+    if(Connectivity().isNotOffline) {
+      Analytics.instance.logSelect(target: "Wellness Center");
+      Navigator.push(context, CupertinoPageRoute(builder: (context) => Covid19WellnessCenter()));
+    } else{
+      AppAlert.showOfflineMessage(context);
+    }
   }
 
   void _onTapLink(String url) {
-    if (AppString.isStringNotEmpty(url)) {
-      if (AppUrl.launchInternal(url)) {
-        Navigator.push(context, CupertinoPageRoute(builder: (context) => WebPanel(url: url)));
-      } else {
-        launch(url);
+    if (Connectivity().isNotOffline) {
+      if (AppString.isStringNotEmpty(url)) {
+        if (AppUrl.launchInternal(url)) {
+          Navigator.push(context, CupertinoPageRoute(builder: (context) => WebPanel(url: url)));
+        } else {
+          launch(url);
+        }
       }
+    } else {
+      AppAlert.showOfflineMessage(context);
     }
   }
 }